Update project title container and add "Projects" text
The commit updates the project title container in `main.js` by adding a line of code to set the text to "Projects". Additionally, the commit also updates the styling of the project title container in `styles.css`. The font size is set to 28px and the background color is forcefully removed.